fix: 修复 anneal_volume_metrics.rs next_index 在 32 位平台(i686)的算术溢出
CI 在 windows-latest x86 (i686-pc-windows-msvc) 上 Build wheels 失败：
  error: this arithmetic operation will overflow
  src/anneal_volume_metrics.rs:115
  attempt to shift right by 32_i32, which would overflow

原因：原表达式  依赖 usize 位宽，
32 位平台上 usize 为 u32，先截断再右移 32 位即溢出。

修复：乘法与移位全程在 u64 上完成，最后才 as usize。
- 64 位平台：逐 bit 等价，行为不变
- 32 位平台：从无法编译变为正确（Lemire 法保证结果 < n，截断无损）

验证：本地 x86_64 cargo check 通过；
i686-unknown-linux-gnu 交叉检查未再出现 arithmetic overflow 错误。

+++ src/anneal_volume_metrics.rs
@@ -112,7 +112,8 @@ impl XorShift64 {
         if n == 0 {
             return 0;
         }
-        ((self.next_u64() >> 32) as u64).wrapping_mul(n as u64) as usize >> 32
+        // 全程在 u64 上完成乘法与移位，避免 32 位平台 usize 截断后 >> 32 溢出
+        ((self.next_u64() >> 32).wrapping_mul(n as u64) >> 32) as usize
     }
 }
